+++ src/routes/character/cash.router.js
@@ -1,6 +1,6 @@
 import express from 'express';
-import authMiddleware from '../middlewares/auth.middleware.js';
-import { prisma } from '../utils/prisma/index.js';
+import authMiddleware from '../../middlewares/auth.middleware.js';
+import { prisma } from '../../utils/prisma/index.js';
 
 const router = express.Router();
