+++ src/routes/character/content/futsal/rank-game.router.js
@@ -1,11 +1,11 @@
 import express from 'express';
-import authMiddleware from '../middlewares/auth.middleware.js';
-import { prisma } from '../utils/prisma/index.js';
+import authMiddleware from '../../../../middlewares/auth.middleware.js';
+import { prisma } from '../../../../utils/prisma/index.js';
 
 const router = express.Router();
 
 // 랭크 풋살 게임 API (JWT 인증)
-router.post('/game-content/futsal/rank-game', authMiddleware, async (req, res, next) => {
+router.post('/character/content/futsal/rank-game', authMiddleware, async (req, res, next) => {
   try {
     const { characterId } = req.character;
     const myCharacter = await prisma.character.findUnique({
